bandziau kazka pacookint su symbol pool

+++ nodes/Symbols/Ankh.cs
@@ -5,7 +5,7 @@ public partial class Ankh : Symbol
 {
 	public override string ImagePath => "res://sprites/ankh.png";
 	public override SymbolType Type => SymbolType.Ankh;
-    public override double SymbolWeight => 1.0;
+    public override int SymbolWeight => 1;
 
 	public Ankh(){}
 }
